requirements.txt
py-ubjson~=0.16
termcolor~=1.1
mypy~=0.910
types-termcolor~=1.1
pyenet
pyee
